requirements.txt
openai
langchain-core
langchain-vectorstores
langchain-openai
langchain-community
streamlit
chromadb
deep-translator
